fix for typo in SempahoreExecutor
