+++ src/test/java/io/github/thunkware/vt/bridge/ExecutorTool8Test.java
@@ -95,9 +95,9 @@ class ExecutorTool8Test {
     }
 
     @Test
-    void testNewSempahoreVirtualExecutorWithAquireTimeout() throws InterruptedException {
+    void testNewSemaphoreVirtualExecutorWithAquireTimeout() throws InterruptedException {
 
-        ExecutorService executor = ExecutorTool.newSempahoreVirtualExecutor(1, Duration.ofMillis(100));
+        ExecutorService executor = ExecutorTool.newSemaphoreVirtualExecutor(1, Duration.ofMillis(100));
         assertThat(executor.isShutdown()).isFalse();
         assertThat(executor.isTerminated()).isFalse();
